Fix indent in gentle_asarray and add Nadia's workaround for the gwcs bug

+++ jwst/combine_1d/combine1d.py
@@ -47,7 +47,10 @@ class InputSpectrumModel:
             to use weight = 1.
         """
 
-        self.wavelength = spec.spec_table.field("wavelength").copy()
+        # We're casting this from dtype('>f8') to dtype('float64') because our version
+        # of gwcs has a bug that prevents it from recognizing big-endian dtypes as
+        # numeric.  Once we upgrade to gwcs 0.11 we can remove the cast.
+        self.wavelength = spec.spec_table.field("wavelength").copy().astype(np.float64)
 
         self.flux = spec.spec_table.field("flux").copy()
         self.error = spec.spec_table.field("error").copy()
